+++ tests/test-save-post.php
@@ -33,10 +33,10 @@ class SavePostTest extends WP_UnitTestCase {
       'post_title' => 'Test Post'
     ]);
     
-    $this->assertNotEmpty(get_field('acfml_post_title_en', $post_id));
-    $this->assertNotEmpty(get_field('acfml_slug_en', $post_id));
-    $this->assertNotEmpty(get_field('acfml_lang_active_en', $post_id));
-    $this->assertEmpty(get_field('acfml_post_title_de', $post_id));
+    $this->assertNotEmpty(get_post_meta($post_id, 'acfml_post_title_en', true));
+    $this->assertNotEmpty(get_post_meta($post_id, 'acfml_slug_en', true));
+    $this->assertNotEmpty(get_post_meta($post_id, 'acfml_lang_active_en', true));
+    $this->assertEmpty(get_post_meta($post_id, 'acfml_post_title_de', true));
 		
 	}
 }